+++ publications_bootstrap/models/__init__.py
@@ -1,8 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from .type import Type
-from .list import List
+from .catalog import Catalog
 from .publication import Publication
-from .customlink import CustomLink
-from .customfile import CustomFile
-from .orderedmodel import OrderedModel
+from .publicationlink import PublicationLink
+from .publicationfile import PublicationFile