+++ publications_bootstrap/templatetags/publication_extras.py
@@ -1,6 +1,4 @@
-__license__ = 'MIT License <http://www.opensource.org/licenses/mit-license.php>'
-__author__ = 'Lucas Theis <lucas@theis.io>'
-__docformat__ = 'epytext'
+# -*- coding: utf-8 -*-
 
 from distutils.version import StrictVersion
 from re import sub
@@ -11,7 +9,7 @@ from django.template.loader import get_template
 from django.utils.html import escape
 from django.utils.safestring import mark_safe
 
-from ..models import Publication, List, Type
+from ..models import Publication, Catalog, Type
 from ..utils import populate
 
 register = Library()
@@ -29,7 +27,7 @@ def render_template(template, request, args):
 
 
 @register.simple_tag(takes_context=True)
-def get_publications(context, template='publications_bootstrap/pages/publications.html'):
+def get_publications(context, template='publications_bootstrap/components/publications.html'):
     """
     Get all publications.
     """
@@ -40,8 +38,7 @@ def get_publications(context, template='publications_bootstrap/pages/publication
     publications = publications.order_by('-year', '-month', '-id')
 
     if not publications:
-        return render_template('publications_bootstrap/components/empty.html',
-                               context['request'], {})
+        return render_template('publications_bootstrap/components/empty.html', context['request'], {})
 
     # load custom links and files
     populate(publications)
@@ -57,27 +54,24 @@ def get_publication(context, p_id):
     try:
         pbl = Publication.objects.get(pk=int(p_id))
 
-        pbl.links = pbl.customlink_set.all()
-        pbl.files = pbl.customfile_set.all()
+        pbl.links = pbl.publicationlink_set.all()
+        pbl.files = pbl.publicationfile_set.all()
 
-        return render_template('publications_bootstrap/components/publication.html',
-                               context['request'],
+        return render_template('publications_bootstrap/components/publication.html', context['request'],
                                {'publication': pbl})
     except Publication.DoesNotExist:
-        return render_template('publications_bootstrap/components/empty.html', context['request'],
-                               {})
+        return render_template('publications_bootstrap/components/empty.html', context['request'], {})
 
 
 @register.simple_tag(takes_context=True)
-def get_publication_list(context, list_title,
-                         template='publications_bootstrap/components/section.html'):
+def get_catalog(context, catalog_title, template='publications_bootstrap/components/section.html'):
     """
-    Get a publication list.
+    Get a publication catalog.
     """
     try:
-        publications_list = List.objects.get(title__iexact=list_title)
+        publications_catalog = Catalog.objects.get(title__iexact=catalog_title)
 
-        publications = publications_list.publication_set.all()
+        publications = publications_catalog.publication_set.all()
         if not publications:
             raise Publication.DoesNotExist
         publications = publications.order_by('-year', '-month', '-id')
@@ -85,17 +79,14 @@ def get_publication_list(context, list_title,
         # load custom links and files
         populate(publications)
 
-        return render_template(template, context['request'],
-                               {'title': list_title, 'publications': publications})
-    except List.DoesNotExist:
+        return render_template(template, context['request'], {'title': catalog_title, 'publications': publications})
+    except Catalog.DoesNotExist:
         return render_template('publications_bootstrap/components/empty.html', context['request'],
                                {'error': True, 'alert':
                                    {'heading': 'Zut!',
-                                    'message': 'There is no such list named "%"'.format(
-                                        list_title)}})
+                                    'message': 'There is no such catalog named "%"'.format(catalog_title)}})
     except Publication.DoesNotExist:
-        return render_template('publications_bootstrap/components/empty.html', context['request'],
-                               {})
+        return render_template('publications_bootstrap/components/empty.html', context['request'], {})
 
 
 @register.filter()