+++ publications_bootstrap/views/__init__.py
@@ -1,10 +1,8 @@
-__license__ = 'MIT License <http://www.opensource.org/licenses/mit-license.php>'
-__author__ = 'Lucas Theis <lucas@theis.io>'
-__docformat__ = 'epytext'
+# -*- coding: utf-8 -*-
 
-from .year import year
-from .author import author
-from .id import id
-from .keyword import keyword
-from .list import list
-from .unapi import unapi
+from .author import by_author
+from .id import by_id
+from .tag import by_tag
+from .catalog import for_catalog
+from .unapi import by_unapi
+from .year import by_year