+++ publications_bootstrap/views/unapi.py
@@ -1,6 +1,4 @@
-__license__ = 'MIT License <http://www.opensource.org/licenses/mit-license.php>'
-__author__ = 'Lucas Theis <lucas@theis.io>'
-__docformat__ = 'epytext'
+# -*- coding: utf-8 -*-
 
 from django.http import HttpResponse
 from django.shortcuts import render
@@ -8,7 +6,7 @@ from django.shortcuts import render
 from ..models import Publication
 
 
-def unapi(request):
+def by_unapi(request):
     """
     This view implements unAPI 1.0 (see http://unapi.info).
     """
@@ -25,53 +23,41 @@ def unapi(request):
 
         except ValueError:
             # invalid id
-            return HttpResponse('\n'.join([
-                '<?xml version="1.0" encoding="UTF-8"?>',
-                '<error>Invalid ID.</error>']),
-                content_type="application/xml",
-                status=404)
+            return HttpResponse('\n'.join(['<?xml version="1.0" encoding="UTF-8"?>', '<error>Invalid ID.</error>']),
+                                content_type="application/xml", status=404)
 
         if format == 'bibtex':
             # return BibTex encoded publication
-            return render(request, 'publications_bootstrap/export/publication.bib', {
-                'publication': publications[0]
-            },
+            return render(request, 'publications_bootstrap/export/publication.bib', {'publication': publications[0]},
                           content_type='text/x-bibtex; charset=UTF-8')
 
         if format == 'mods':
             # return MODS encoded publication
-            return render(request, 'publications_bootstrap/export/publications.mods', {
-                'publications': publications
-            },
+            return render(request, 'publications_bootstrap/export/publications.mods', {'publications': publications},
                           content_type='application/xml; charset=UTF-8')
 
         if format == 'ris':
             # return MODS encoded publication
-            return render(request, 'publications_bootstrap/export/publications.ris', {
-                'publications': publications
-            },
+            return render(request, 'publications_bootstrap/export/publications.ris', {'publications': publications},
                           content_type='application/x-research-info-systems; charset=UTF-8')
 
         # invalid format
-        return HttpResponse('\n'.join([
-            '<?xml version="1.0" encoding="UTF-8"?>',
-            '<error>Invalid format.</error>']),
-            content_type="application/xml",
-            status=406)
+        return HttpResponse('\n'.join(['<?xml version="1.0" encoding="UTF-8"?>', '<error>Invalid format.</error>']),
+                            content_type="application/xml", status=406)
 
     if id is not None:
-        return HttpResponse('\n'.join([
-            '<?xml version="1.0" encoding="UTF-8"?>',
-            '<formats id="{0}">'.format(id),
-            '<format name="bibtex" type="text/x-bibtex" />',
-            '<format name="ris" type="application/x-research-info-systems" />',
-            '<format name="mods" type="application/xml" />',
-            '</formats>']), content_type="application/xml")
+        return HttpResponse('\n'.join(['<?xml version="1.0" encoding="UTF-8"?>',
+                                       '<formats id="{0}">'.format(id),
+                                       '<format name="bibtex" type="text/x-bibtex" />',
+                                       '<format name="ris" type="application/x-research-info-systems" />',
+                                       '<format name="mods" type="application/xml" />',
+                                       '</formats>']),
+                            content_type="application/xml")
 
-    return HttpResponse('\n'.join([
-        '<?xml version="1.0" encoding="UTF-8"?>',
-        '<formats>',
-        '<format name="bibtex" type="text/x-bibtex" />',
-        '<format name="ris" type="application/x-research-info-systems" />',
-        '<format name="mods" type="application/xml" />',
-        '</formats>']), content_type="application/xml")
+    return HttpResponse('\n'.join(['<?xml version="1.0" encoding="UTF-8"?>',
+                                   '<formats>',
+                                   '<format name="bibtex" type="text/x-bibtex" />',
+                                   '<format name="ris" type="application/x-research-info-systems" />',
+                                   '<format name="mods" type="application/xml" />',
+                                   '</formats>']),
+                        content_type="application/xml")